second round of cleanup
