+++ openmls/src/group/core_group/mod.rs
@@ -160,7 +160,7 @@ pub(crate) struct StagedCoreWelcome {
 
 #[derive(Debug, Serialize, Deserialize)]
 #[cfg_attr(any(test, feature = "test-utils"), derive(Clone, PartialEq))]
-pub struct CoreGroup {
+pub(crate) struct CoreGroup {
     public_group: PublicGroup,
     group_epoch_secrets: GroupEpochSecrets,
     own_leaf_index: LeafNodeIndex,
@@ -1196,6 +1196,7 @@ impl CoreGroup {
 }
 
 // Test and test-utils functions
+#[cfg_attr(all(not(test), feature = "test-utils"), allow(dead_code))]
 #[cfg(any(feature = "test-utils", test))]
 impl CoreGroup {
     pub(crate) fn context_mut(&mut self) -> &mut GroupContext {
@@ -1210,7 +1211,7 @@ impl CoreGroup {
         println!("{}: {}", message, self.public_group().export_ratchet_tree());
     }
 
-    pub fn resumption_psk_store(&self) -> &ResumptionPskStore {
+    pub(crate) fn resumption_psk_store(&self) -> &ResumptionPskStore {
         &self.resumption_psk_store
     }
 }
