+++ openmls/src/group/mls_group/mod.rs
@@ -467,8 +467,8 @@ impl MlsGroup {
     }
 
     /// Returns the underlying [CoreGroup].
-    #[cfg(any(test, feature = "test-utils"))]
-    pub fn group(&self) -> &CoreGroup {
+    #[cfg(test)]
+    pub(crate) fn group(&self) -> &CoreGroup {
         &self.group
     }
