+++ openmls/src/group/public_group/diff.rs
@@ -239,7 +239,7 @@ impl<'a> PublicGroupDiff<'a> {
 /// modified. Its only use is to merge it into the original [`PublicGroup`].
 #[derive(Debug, Serialize, Deserialize)]
 #[cfg_attr(any(test, feature = "test-utils"), derive(Clone, PartialEq))]
-pub struct StagedPublicGroupDiff {
+pub(crate) struct StagedPublicGroupDiff {
     pub(super) staged_diff: StagedTreeSyncDiff,
     pub(super) group_context: GroupContext,
     pub(super) interim_transcript_hash: Vec<u8>,
