+++ openmls/src/messages/proposals.rs
@@ -478,7 +478,7 @@ impl GroupContextExtensionProposal {
     }
 
     /// Get the extensions of the proposal
-    pub fn extensions(&self) -> &Extensions {
+    pub(crate) fn extensions(&self) -> &Extensions {
         &self.extensions
     }
 }
