+++ openmls/src/schedule/mod.rs
@@ -1225,7 +1225,7 @@ impl EpochSecrets {
 
 #[derive(Serialize, Deserialize)]
 #[cfg_attr(any(test, feature = "test-utils"), derive(Clone, Debug, PartialEq))]
-pub struct GroupEpochSecrets {
+pub(crate) struct GroupEpochSecrets {
     init_secret: InitSecret,
     exporter_secret: ExporterSecret,
     epoch_authenticator: EpochAuthenticator,
