+++ openmls/src/schedule/psk.rs
@@ -491,7 +491,7 @@ pub mod store {
     /// This is where the resumption PSKs are kept in a rollover list.
     #[derive(Debug, Serialize, Deserialize)]
     #[cfg_attr(any(test, feature = "test-utils"), derive(Clone, PartialEq))]
-    pub struct ResumptionPskStore {
+    pub(crate) struct ResumptionPskStore {
         max_number_of_secrets: usize,
         resumption_psk: Vec<(GroupEpoch, ResumptionPskSecret)>,
         cursor: usize,
@@ -533,9 +533,9 @@ pub mod store {
         }
     }
 
-    #[cfg(any(feature = "test-utils", test))]
+    #[cfg(test)]
     impl ResumptionPskStore {
-        pub fn cursor(&self) -> usize {
+        pub(crate) fn cursor(&self) -> usize {
             self.cursor
         }
     }
