+++ openmls/src/treesync/tests_and_kats/kats.rs
@@ -1,6 +1,8 @@
 // FIXME: 624 The tree kem test vectors have to be regenerated after #541.
 // pub mod kat_tree_kem;
 #[cfg(test)]
+pub mod kat_storage_stability;
+#[cfg(test)]
 pub mod kat_tree_operations;
 #[cfg(test)]
 pub mod kat_tree_validation;
